+++ rltk/visibility.hpp
@@ -32,7 +32,7 @@ void internal_2d_sweep(const location_t_ &position, const int &range, std::funct
 	
 	line_func_cancellable(start_x, start_y, end_x, end_y, [&blocked, &is_opaque, &set_visible, &range, &position] (int X, int Y) {
 		if (blocked) return false;
-		float distance = distance2d(position.x, position.y, X, Y);
+		float distance = distance2d(static_cast<int>(position.x), static_cast<int>(position.y), X, Y);
 		if (distance <= range) {
 			location_t_ pos = navigator_t::get_xy(X,Y);
 			if (!blocked) set_visible(pos);
